Migrate card page from /docs/components/card to /components/card
Port all examples (Image Card, Stats Cards, Profile Card, Login Form)
and CardImage API reference to the new reference page. Remove old page,
update routes, navigation links, and E2E tests.
